@@ -12,7 +12,7 @@ class ChatListNotifier extends StateNotifier<ChatListState> {
 
   void _loadUp() async {
     state = asyncChats.when(
-      data: (rooms) => ChatListState.data(chats: rooms),
+      data: (rooms) => ChatListState.data(chats: [...rooms]),
       error: (e, s) => ChatListState.error(e.toString()),
       loading: () => const ChatListState.loading(),
     );
