@@ -13,6 +13,7 @@ class LocalSyncState {
 class SyncNotifier extends StateNotifier<LocalSyncState> {
   late SyncState syncState;
   late Stream<bool>? syncPoller;
+  late Stream<String>? notifications;
   late Stream<bool>? errorPoller;
   late Ref ref;
 
@@ -20,8 +21,9 @@ class SyncNotifier extends StateNotifier<LocalSyncState> {
     startSync(client, ref);
   }
 
+  Stream<String>? get notificationsStream => notifications;
+
   Future<void> startSync(Client client, Ref ref) async {
-    // Get.put(ReceiptController(client: state!));
     // on release we have a really weird behavior, where, if we schedule
     // any async call in rust too early, they just pend forever. this
     // hack unfortunately means we have two wait a bit but that means
